+++ CHANGELOG.md
@@ -3,6 +3,7 @@ TBD
 Unreleased changes. Release notes have not yet been written.
 
 * add extern crate memchr into memchr bench
+* add arm support to memset()
 
 0.1.1 (2021-05-14)
 =====